+++ packages/client/hmi-client/src/components/operator/tera-operator-header.vue
