+++ packages/client/hmi-client/src/workflow/ops/calibrate-julia/tera-calibrate-node-julia.vue
@@ -15,7 +15,7 @@
 
 <script setup lang="ts">
 import { computed } from 'vue';
-import TeraOperatorPlaceholder from '@/workflow/operator/tera-operator-placeholder.vue';
+import TeraOperatorPlaceholder from '@/components/operator/tera-operator-placeholder.vue';
 import { WorkflowNode } from '@/types/workflow';
 import Button from 'primevue/button';
 import { CalibrationOperationStateJulia } from './calibrate-operation';
