+++ packages/client/hmi-client/src/workflow/ops/funman/tera-funman-node.vue
@@ -17,7 +17,7 @@
 </template>
 
 <script setup lang="ts">
-import TeraOperatorPlaceholder from '@/workflow/operator/tera-operator-placeholder.vue';
+import TeraOperatorPlaceholder from '@/components/operator/tera-operator-placeholder.vue';
 import { WorkflowNode } from '@/types/workflow';
 import { FunmanOperationState } from '@/workflow/ops/funman/funman-operation';
 import Button from 'primevue/button';
