+++ packages/client/hmi-client/src/workflow/ops/model-optimize/tera-model-optimize-node.vue
@@ -17,7 +17,7 @@
 </template>
 
 <script setup lang="ts">
-import TeraOperatorPlaceholder from '@/workflow/operator/tera-operator-placeholder.vue';
+import TeraOperatorPlaceholder from '@/components/operator/tera-operator-placeholder.vue';
 import { WorkflowNode } from '@/types/workflow';
 import Button from 'primevue/button';
 import { ModelOptimizeOperationState } from './model-optimize-operation';
